V6

+++ src/main/java/org/sid/gestion_v/service/ServiceImpl/UtilisateurServiceImpl.java
@@ -1,7 +1,9 @@
 package org.sid.gestion_v.service.ServiceImpl;
 
 import lombok.RequiredArgsConstructor;
+import org.sid.gestion_v.entities.Affectation;
 import org.sid.gestion_v.entities.Utilisateur;
+import org.sid.gestion_v.repository.AffectationRepository;
 import org.sid.gestion_v.repository.UtilisateurRepository;
 import org.sid.gestion_v.service.UtilisateurService;
 import org.springframework.stereotype.Service;
@@ -12,6 +14,8 @@ import java.util.List;
 @RequiredArgsConstructor
 public class UtilisateurServiceImpl implements UtilisateurService {
     private final UtilisateurRepository utilisateurRepository;
+    private final AffectationRepository affectationRepository;
+
 
     @Override
     public Utilisateur save(Utilisateur utilisateur) {
@@ -42,6 +46,11 @@ public class UtilisateurServiceImpl implements UtilisateurService {
 
     @Override
     public void deleteUtilisateur(Long id) {
+        List<Affectation> affectations = affectationRepository.findByUtilisateurId(id);
+        if (!affectations.isEmpty()) {
+            throw new RuntimeException("Impossible de supprimer l’utilisateur : il est lié à des affectations.");
+        }
         utilisateurRepository.deleteById(id);
     }
+
 }
